Move remove_experiment_parts
